1.移除下限告警功能，使扭力超标的告警变得及时.2.给小灯初始值亮50个，扭力值存在负值

+++ User/WorkMode.h
@@ -27,7 +27,7 @@ extern uint8_t CurrentIndex2;
 extern uint8_t LastIndex2;
 extern uint8_t Ceiling_Value;
 extern uint8_t Floor_Value;
-extern int offset;
+extern float offset;
 /**********************************************************************
 fun extern
 ***********************************************************************/
